Add new RevisionSnapshot
In order to isolate Revision processes, a new type of snapshot for the Qbeast structures is created.
QbeastSnapshot would be only in charge of grouping the different RevisionSnapshots that can be found in the proper delta Snapshot.
Analyze, Optimize and Write operations would affect only one RevisionSnapshot,
and Read operation needs to query all available RevisionSnapshots.