finsihed using new Rev class, still a few bugs to fix
I have change the way the components are wired togheter and started
tu use the new Revision* classes in the code. I had a some problem
with the JSON serialization of the object, that should be reviewed.
There are a few tests that still fails. I created the IndexStatus as
a trait but it should be probably be a plain case class, and use the
DeltaIndexStatus to generate it ( and it should not inherit from
IndexStatus otherwise it gives problems in serializations).

There are also plenty of TODOs to check, code to review, and design
decision to discuss.