Using Dataset as often that possible
As we begin to handle a significant amount of metadata and transfer
everything to the driver, we start encountering issues.
One such problem stems from an outdated design approach where we
aimed to keep the core of Qbeast independent from Spark. However,
this approach now seems less sensible. If we aim to support different
query engines in the future, it would be more efficient to rewrite
the core classes in the respective languages of those engines.

So what I've done is:
1. Move all classes from the core package to the spark one (only one package)
2. Change some API, so we return Dataset instead of IISeq
3. Change the code to rely as much as possible to Spark to do the
   computation